Fix warning and update pod spec file.

+++ MessageDisplayKit/Classes/Categorys/UIImage/UIImage+Resize.m
@@ -132,7 +132,7 @@
             break;
             
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", contentMode];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %ld", (long)contentMode];
     }
     
     CGSize newSize = CGSizeMake(self.size.width * ratio, self.size.height * ratio);
